Fix shrink_to_fit; add examples, enable valgrind and other tests
This commit fixes a bug in shrink_in_place, which ideally shrinks an allocation to
the size requested, but can shrink it to a slightly larger size (which an
_excess variant would return). However, previously, if the size after shrinking
was equal or larger than the desired size, the function would return an error,
only returning success if the allocation was shrunk to a size smaller than the
requested one, which is something that just cannot happen (otherwise that's a
bug in jemalloc).

So now it at least has the "right" behavior, but the API is still partially
borked, because the error condition with the current API cannot really ever
happen, unless the allocator has a bug or it does not support shrinking in place
at all (in which case it could always return the error, but it could also always
return success...).

This commit adds:

* examples in the example's directory
  * runs the examples under valgrind on Linux with jemallocator as the global allocator

  * runs the examples under address sanitizer on Linux when jemallocator is not
    the global allocator but still used (the sanitizers replace the global
    allocator anyways, so other situations do not make much sense)

  * compiles the examples with thread and memory sanitizer on linux when
  jemallocator is not the global allocator but still used

* test that the no_std jemallocator builds do not pull any symbols from std

* runs jemalloc's tests on more platforms to make sure that jemalloc-sys works
  correctly

There are some pass-by cleanups/path canonicalization in the build.rs. I was
trying to get the windows build bots up and running, and the path
canonicalizations where helpfull there, but I've decided to submit Windows
support in a different PR because that's not finished yet.